+++ packages/@react-enhanced/translate/shim.d.ts
@@ -1,0 +1,9 @@
+declare namespace i18n {
+  export interface Translations {
+    en: import('./src/types.js').Translation
+    zh: import('./src/types.js').Translation
+  }
+
+  // eslint-disable-next-line no-restricted-syntax
+  export type I18n = Translations
+}