+++ CHANGELOG.md
@@ -5,5 +5,10 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - [@react-enhanced/eslint-plugin](./packages/@react-enhanced/eslint-plugin/CHANGELOG.md)
 - [@react-enhanced/hooks](./packages/@react-enhanced/hooks/CHANGELOG.md)
+- [@react-enhanced/plugins](./packages/@react-enhanced/plugins/CHANGELOG.md)
+- [@react-enhanced/shared](./packages/@react-enhanced/shared/CHANGELOG.md)
+- [@react-enhanced/translate](./packages/@react-enhanced/translate/CHANGELOG.md)
+- [@react-enhanced/types](./packages/@react-enhanced/types/CHANGELOG.md)
+- [@react-enhanced/utils](./packages/@react-enhanced/utils/CHANGELOG.md)
 - [react-qrcode](./packages/react-qrcode/CHANGELOG.md)
 - [react-qrious](./packages/react-qrious/CHANGELOG.md)